+++ apps/meteor/app/slashcommands-topic/client/topic.ts
@@ -1,9 +1,9 @@
 import type { SlashCommandCallbackParams } from '@rocket.chat/core-typings';
 import { clientCallbacks } from '@rocket.chat/ui-client';
 
+import { hasPermission } from '../../../client/lib/authorization';
 import { dispatchToastMessage } from '../../../client/lib/toast';
 import { Rooms } from '../../../client/stores';
-import { hasPermission } from '../../authorization/client';
 import { sdk } from '../../utils/client/lib/SDKClient';
 import { slashCommands } from '../../utils/client/slashCommand';
