+++ apps/meteor/client/lib/authorization/hasPermission.ts
