+++ apps/meteor/client/lib/authorization/hasRole.ts
