+++ apps/meteor/client/views/marketplace/sidebarItems.tsx
@@ -1,5 +1,5 @@
 import MarketplaceRequestBadge from './components/MarketplaceRequestBadge';
-import { hasAtLeastOnePermission, hasPermission } from '../../../app/authorization/client';
+import { hasAtLeastOnePermission, hasPermission } from '../../lib/authorization';
 import { createSidebarItems } from '../../lib/createSidebarItems';
 import { links } from '../../lib/links';
