+++ apps/meteor/client/views/omnichannel/contactInfo/EditContactInfo.tsx
@@ -22,9 +22,9 @@ import { useTranslation } from 'react-i18next';
 import AdvancedContactModal from './AdvancedContactModal';
 import { useCreateContact } from './hooks/useCreateContact';
 import { useEditContact } from './hooks/useEditContact';
-import { hasAtLeastOnePermission } from '../../../../app/authorization/client';
 import { useFormSubmitWithDirtyCheck } from '../../../hooks/useFormSubmitWithDirtyCheck';
 import { useHasLicenseModule } from '../../../hooks/useHasLicenseModule';
+import { hasAtLeastOnePermission } from '../../../lib/authorization';
 import { omnichannelQueryKeys } from '../../../lib/queryKeys';
 import { ContactManagerInput } from '../additionalForms';
 import { useCustomFieldsMetadata } from '../directory/hooks/useCustomFieldsMetadata';
