+++ apps/meteor/client/views/omnichannel/directory/chats/ChatInfo/RoomEdit/RoomEdit.tsx
@@ -6,7 +6,7 @@ import { useQueryClient } from '@tanstack/react-query';
 import { useCallback, useId } from 'react';
 import { useController, useForm } from 'react-hook-form';
 
-import { hasAtLeastOnePermission } from '../../../../../../../app/authorization/client';
+import { hasAtLeastOnePermission } from '../../../../../../lib/authorization';
 import { roomsQueryKeys } from '../../../../../../lib/queryKeys';
 import { SlaPoliciesSelect, PrioritiesSelect } from '../../../../additionalForms';
 import Tags from '../../../../components/Tags';
