+++ apps/meteor/client/views/room/providers/ComposerPopupProvider.tsx
@@ -9,9 +9,10 @@ import { useMemo, useState } from 'react';
 import type { ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { hasAtLeastOnePermission } from '../../../../app/authorization/client';
 import { emoji } from '../../../../app/emoji/client';
 import { slashCommands } from '../../../../app/utils/client';
+import { hasAtLeastOnePermission } from '../../../lib/authorization';
+import { pipe } from '../../../lib/cachedStores/pipe';
 import { cannedResponsesQueryKeys } from '../../../lib/queryKeys';
 import { Messages, Subscriptions } from '../../../stores';
 import ComposerBoxPopupCannedResponse from '../composer/ComposerBoxPopupCannedResponse';
@@ -27,7 +28,6 @@ import type { ComposerPopupContextValue, ComposerPopupOption } from '../contexts
 import { ComposerPopupContext, createMessageBoxPopupConfig } from '../contexts/ComposerPopupContext';
 import useCannedResponsesQuery from './hooks/useCannedResponsesQuery';
 import { normalizeUsername } from '../../../../lib/utils/normalizeUsername';
-import { pipe } from '../../../lib/cachedStores/pipe';
 
 const getToneRegexes = () => {
 	const exactFinalTone = new RegExp('^tone[1-5](?:-[1-5])?:*$');
